transaction

+++ src/main/java/ru/mariamaximova/bl1/application/comment/service/impl/CommentServiceImpl.java
@@ -1,6 +1,7 @@
 package ru.mariamaximova.bl1.application.comment.service.impl;
 
 import lombok.AllArgsConstructor;
+import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -17,6 +18,7 @@ import ru.mariamaximova.bl1.application.fiml.domain.FilmRepository;
 import ru.mariamaximova.bl1.application.rating.service.RatingService;
 import ru.mariamaximova.bl1.error.ErrorDescription;
 
+import javax.transaction.UserTransaction;
 import java.util.Comparator;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -34,6 +36,8 @@ public class CommentServiceImpl implements CommentService {
 
     private final RatingService ratingService;
 
+    private final UserTransaction userTransaction;
+
     @Override
     public List<ResponseCommentDto> getComments(Long filmId) {
         log.info("start getComment({})", filmId);
@@ -44,34 +48,59 @@ public class CommentServiceImpl implements CommentService {
 
     @Override
     @Transactional
+    @SneakyThrows
     public void saveComment(Long filmId, Long customerId, CommentDto commentDto) {
         log.info("start saveComment({}, {}, {})", filmId, customerId, commentDto);
+
+        userTransaction.begin();
         Comment comment = commentRepository.getByFilmIdAndCustomerId(filmRepository.getById(filmId),
                 customerRepository.getById(customerId));
         if (ObjectUtils.isEmpty(comment) || !ObjectUtils.isEmpty(commentDto.getId()) &&
                 comment.getId().equals(commentDto.getId())) {
             commentRepository.save(convertToComment(filmId, customerId, commentDto));
+            userTransaction.commit();
+            ratingService.saveRating(filmId, customerId, commentDto.getRating());
         } else {
+            userTransaction.rollback();
             log.info("Error save uniq");
             throw ErrorDescription.SAVE_COMMENT_ERROR_UNIQ.exception();
         }
-        ratingService.saveRating(filmId, customerId, commentDto.getRating());
+
         log.info("complete save");
     }
 
 
     @Override
     @Transactional
+    @SneakyThrows
     public void deleteComment(Long comment_id) {
-       commentRepository.deleteById(comment_id);
-       ratingService.deleteRating(comment_id);
+        userTransaction.begin();
+        if (!ObjectUtils.isEmpty(commentRepository.findById(comment_id))) {
+            commentRepository.deleteById(comment_id);
+            ratingService.deleteRating(comment_id);
+            userTransaction.commit();
+        } else {
+            userTransaction.rollback();
+            log.info("Error save uniq");
+            ErrorDescription.COMMENT_NOT_FOUND.throwException();
+        }
+
     }
 
     @Override
     @Transactional
+    @SneakyThrows
     public void updateStatusComment(Long comment_id) {
-        commentRepository.getCommentById(comment_id).set_active(true);
-        ratingService.updateStatusRating(comment_id);
+        userTransaction.begin();
+        if (!ObjectUtils.isEmpty(commentRepository.findById(comment_id))) {
+            commentRepository.getById(comment_id).set_active(true);
+            ratingService.updateStatusRating(comment_id);
+            userTransaction.commit();
+        } else {
+            userTransaction.rollback();
+            log.info("Error save uniq");
+            throw ErrorDescription.COMMENT_NOT_FOUND.exception();
+        }
     }
 
     private ResponseCommentDto convertToCommentDto(Comment comment) {
